requirements.txt
librosa
pandas
numpy
matplotlib
seaborn
sklearn
tensorflow
tqdm
pyYAML
-e .